requirements.txt
openai
PyPDF2
langchain
tqdm
python-dotenv
faiss-cpu
huggingface-hub